pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-code"
version = "0.1.0"
description = "AI-powered code analysis, refactoring, and language conversion CLI tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "openai>=2.0",
    "python-dotenv>=1.0",
    "typer>=0.9",
    "rich>=13.0",
    "tqdm>=4.60",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "mypy>=1.8"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = ["tests/"]

[tool.coverage.run]
source = ["ai_code"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 60
